// src/redux/web3/web3.action.types.ts
export enum ActionType {
  INIT_WEB3_PENDING = "INIT_WEB3_PENDING",
  INIT_WEB3_SUCCESS = "INIT_WEB3_SUCCESS",
  INIT_WEB3_FAIL = "INIT_WEB3_FAIL",
  GET_ACCOUNT_PENDING = "GET_ACCOUNT_PENDING",
  GET_ACCOUNT_SUCCESS = "GET_ACCOUNT_SUCCESS",
  GET_ACCOUNT_FAIL = "GET_ACCOUNT_FAIL",
}

export type Action = {
  type: ActionType,
  payload?: any
}